+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4] - Unreleased
+## [1.4] - 2024-10-02
 
 ### Added
 
 - Add ``row_wise_max`` function
 - Add ``row_wise_add`` function
 - Add ``row_wise_sub`` wrapper function
+- Expose functions in ``__init__.py``
 
 ### Changed